Remove NodeExt + PackedSceneExt, use inherent impl instead
Improves discoverability, since the extended methods now appear in the original class doc.
Also simplifies handling since no external trait needs to be imported. From a user perspective,
it's not really relevant if a method is generated or manually written; the border is already
fluent with "extra" methods like _ex() default parameters, or Object::try_call().
