+++ godot-core/src/engine/manual_extensions.rs
@@ -1,0 +1,74 @@
+/*
+ * Copyright (c) godot-rust; Bromeon and contributors.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+use crate::builtin::NodePath;
+use crate::engine::{Node, PackedScene};
+use crate::obj::{Gd, Inherits};
+
+/// Manual extensions for the `Node` class.
+impl Node {
+    /// ⚠️ Retrieves the node at path `path`, panicking if not found or bad type.
+    ///
+    /// # Panics
+    /// If the node is not found, or if it does not have type `T` or inherited.
+    pub fn get_node_as<T>(&self, path: impl Into<NodePath>) -> Gd<T>
+    where
+        T: Inherits<Node>,
+    {
+        let path = path.into();
+        let copy = path.clone(); // TODO avoid copy
+
+        self.try_get_node_as(path).unwrap_or_else(|| {
+            panic!(
+                "There is no node of type {ty} at path `{copy}`",
+                ty = T::class_name()
+            )
+        })
+    }
+
+    /// Retrieves the node at path `path` (fallible).
+    ///
+    /// If the node is not found, or if it does not have type `T` or inherited,
+    /// `None` will be returned.
+    pub fn try_get_node_as<T>(&self, path: impl Into<NodePath>) -> Option<Gd<T>>
+    where
+        T: Inherits<Node>,
+    {
+        let path = path.into();
+
+        // TODO differentiate errors (not found, bad type) with Result
+        self.get_node_or_null(path)
+            .and_then(|node| node.try_cast::<T>().ok())
+    }
+}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+/// Manual extensions for the `PackedScene` class.
+impl PackedScene {
+    /// ⚠️ Instantiates the scene as type `T`, panicking if not found or bad type.
+    ///
+    /// # Panics
+    /// If the scene is not type `T` or inherited.
+    pub fn instantiate_as<T>(&self) -> Gd<T>
+    where
+        T: Inherits<Node>,
+    {
+        self.try_instantiate_as::<T>()
+            .unwrap_or_else(|| panic!("Failed to instantiate {to}", to = T::class_name()))
+    }
+
+    /// Instantiates the scene as type `T` (fallible).
+    ///
+    /// If the scene is not type `T` or inherited.
+    pub fn try_instantiate_as<T>(&self) -> Option<Gd<T>>
+    where
+        T: Inherits<Node>,
+    {
+        self.instantiate().and_then(|gd| gd.try_cast::<T>().ok())
+    }
+}
