+++ godot/src/prelude.rs
@@ -17,15 +17,13 @@ pub use super::builtin::meta::{FromGodot, ToGodot};
 pub use super::engine::{
     load, try_load, utilities, AudioStreamPlayer, Camera2D, Camera3D, GFile, IAudioStreamPlayer,
     ICamera2D, ICamera3D, INode, INode2D, INode3D, IObject, IPackedScene, IRefCounted, IResource,
-    ISceneTree, Input, Node, Node2D, Node3D, Object, PackedScene, PackedSceneExt, RefCounted,
-    Resource, SceneTree,
+    ISceneTree, Input, Node, Node2D, Node3D, Object, PackedScene, RefCounted, Resource, SceneTree,
 };
 pub use super::init::{gdextension, ExtensionLibrary, InitLevel};
 pub use super::log::*;
 pub use super::obj::{Base, Gd, GdMut, GdRef, GodotClass, Inherits, InstanceId, OnReady};
 
 // Make trait methods available.
-pub use super::engine::NodeExt as _;
 pub use super::obj::EngineBitfield as _;
 pub use super::obj::EngineEnum as _;
 pub use super::obj::NewAlloc as _;
