+++ itest/rust/src/engine_tests/node_test.rs
@@ -8,7 +8,7 @@
 use std::str::FromStr;
 
 use godot::builtin::{NodePath, Variant};
-use godot::engine::{global, Node, Node3D, NodeExt, PackedScene, SceneTree};
+use godot::engine::{global, Node, Node3D, PackedScene, SceneTree};
 use godot::obj::{NewAlloc, NewGd};
 
 use crate::framework::{itest, TestContext};
